Added ExcelExportTest
ExcelExportTest shows how a Excel spreadsheet can be dynamically created using the OpenXML sdk.

You can pass a dynamic list of objects to it and it will create a row for each object in the list and put the value of it's properties in the cells (in the order the properties are declared).